stripe/save_to_postgres/requirements.txt
aws-psycopg2==1.2.1
pkg-resources==0.0.0
psycopg2-binary==2.8.5